00-bootcamp-project/pyproject.toml
[tool.poetry]
name = "apis"
version = "0.1.0"
description = ""
authors = ["zkan <zkan.cs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
pysftp = "^0.2.9"
psycopg2 = "^2.9.6"
pandas = "^2.0.1"
google-cloud-bigquery = "^3.10.0"
pyarrow = "^12.0.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
